app/requirements.txt
pymongo==4.10
pytz==2024.2
streamlit
pyyaml